Fix  file length max (#71)
* broke up mud_game_live.ex

* broke up map_selection_live.ex into two files

* npcs.ex and map.ex broken up

* formatting

* test: add comprehensive tests for MudGameLive with 100% coverage


* aider dumb

* changed threshold  to 19% so i can pull this and everyone can get to work

---------
