+++ lib/shard_web/live/admin_live/map.ex
@@ -1,10 +1,13 @@
 defmodule ShardWeb.AdminLive.Map do
+  @moduledoc """
+  Main LiveView for map management.
+  Delegates event handling to MapHandlers and imports UI components from MapComponents.
+  """
   use ShardWeb, :live_view
 
   alias Shard.Map
-  alias Shard.Map.{Room, Door}
-  alias Shard.Repo
-  alias Shard.AI
+  alias ShardWeb.AdminLive.{MapHandlers, MapComponents}
+  import MapComponents
 
   @impl true
   def mount(_params, _session, socket) do
@@ -34,348 +37,6 @@ defmodule ShardWeb.AdminLive.Map do
   defp apply_action(socket, :index, _params) do
     socket
     |> assign(:page_title, "Map Management")
-  end
-
-  @impl true
-  def handle_event("change_tab", %{"tab" => tab}, socket) do
-    {:noreply, assign(socket, :tab, tab)}
-  end
-
-  # Room events
-  def handle_event("new_room", _params, socket) do
-    changeset = Map.change_room(%Room{})
-    {:noreply, assign(socket, :changeset, changeset) |> assign(:editing, :room)}
-  end
-
-  def handle_event("edit_room", %{"id" => id}, socket) do
-    room = Map.get_room!(id)
-    changeset = Map.change_room(room)
-    {:noreply, assign(socket, :changeset, changeset) |> assign(:editing, :room)}
-  end
-
-  def handle_event("view_room", %{"id" => id}, socket) do
-    room = Map.get_room!(id)
-    doors_from = Map.get_doors_from_room(room.id) |> Repo.preload(:to_room)
-    doors_to = Map.get_doors_to_room(room.id) |> Repo.preload(:from_room)
-
-    {:noreply,
-     socket
-     |> assign(:viewing, room)
-     |> assign(:doors_from, doors_from)
-     |> assign(:doors_to, doors_to)
-     |> assign(:changeset, Map.change_room(room))
-     |> assign(:tab, "room_details")}
-  end
-
-  def handle_event("delete_room", %{"id" => id}, socket) do
-    room = Map.get_room!(id)
-    {:ok, _} = Map.delete_room(room)
-
-    rooms = Map.list_rooms()
-    doors = Map.list_doors()
-
-    {:noreply,
-     socket
-     |> assign(:rooms, rooms)
-     |> assign(:doors, doors)
-     |> put_flash(:info, "Room deleted successfully")}
-  end
-
-  def handle_event("validate_room", %{"room" => room_params}, socket) do
-    changeset =
-      if socket.assigns.editing == :room && socket.assigns.changeset.data.id do
-        Map.change_room(socket.assigns.changeset.data, room_params)
-      else
-        Map.change_room(%Room{}, room_params)
-      end
-      |> Map.put_action(:validate)
-
-    {:noreply, assign(socket, :changeset, changeset)}
-  end
-
-  def handle_event("save_room", %{"room" => room_params}, socket) do
-    case save_room(socket, room_params) do
-      {:ok, socket} -> {:noreply, socket}
-      {:error, socket} -> {:noreply, socket}
-    end
-  end
-
-  def handle_event("apply_and_save", %{"room" => room_params}, socket) do
-    case Map.update_room(socket.assigns.viewing, room_params) do
-      {:ok, updated_room} ->
-        rooms = Map.list_rooms()
-
-        {:noreply,
-         socket
-         |> assign(:rooms, rooms)
-         |> assign(:viewing, updated_room)
-         |> assign(:changeset, Map.change_room(updated_room))
-         |> put_flash(:info, "Room updated successfully")}
-
-      {:error, changeset} ->
-        {:noreply, assign(socket, :changeset, changeset)}
-    end
-  end
-
-  def handle_event("generate_description", _params, socket) do
-    zone_description = "A dark and mysterious forest."
-
-    # Use the currently viewed room as the context for finding adjacent rooms.
-    surrounding_rooms =
-      if room = socket.assigns.viewing do
-        Map.get_adjacent_rooms(room.id)
-      else
-        []
-      end
-
-    case AI.generate_room_description(zone_description, surrounding_rooms) do
-      {:ok, description} ->
-        changeset =
-          Ecto.Changeset.put_change(socket.assigns.changeset, :description, description)
-
-        {:noreply, assign(socket, :changeset, changeset)}
-
-      {:error, reason} ->
-        {:noreply, put_flash(socket, :error, "Failed to generate description: #{reason}")}
-    end
-  end
-
-  def handle_event("cancel_room", _params, socket) do
-    {:noreply, assign(socket, :editing, nil) |> assign(:changeset, nil)}
-  end
-
-  def handle_event("back_to_rooms", _params, socket) do
-    {:noreply, assign(socket, :tab, "rooms")}
-  end
-
-  # Door events
-  def handle_event("new_door", _params, socket) do
-    changeset = Map.change_door(%Door{})
-    {:noreply, assign(socket, :changeset, changeset) |> assign(:editing, :door)}
-  end
-
-  def handle_event("edit_door", %{"id" => id}, socket) do
-    door = Map.get_door!(id)
-    changeset = Map.change_door(door)
-    {:noreply, assign(socket, :changeset, changeset) |> assign(:editing, :door)}
-  end
-
-  def handle_event("delete_door", %{"id" => id}, socket) do
-    door = Map.get_door!(id)
-    {:ok, _} = Map.delete_door(door)
-
-    rooms = Map.list_rooms()
-    doors = Map.list_doors()
-
-    {:noreply,
-     socket
-     |> assign(:rooms, rooms)
-     |> assign(:doors, doors)
-     |> put_flash(:info, "Door deleted successfully")}
-  end
-
-  def handle_event("validate_door", %{"door" => door_params}, socket) do
-    changeset =
-      if socket.assigns.editing == :door && socket.assigns.changeset.data.id do
-        Map.change_door(socket.assigns.changeset.data, door_params)
-      else
-        Map.change_door(%Door{}, door_params)
-      end
-      |> Map.put_action(:validate)
-
-    {:noreply, assign(socket, :changeset, changeset)}
-  end
-
-  def handle_event("save_door", %{"door" => door_params}, socket) do
-    case save_door(socket, door_params) do
-      {:ok, socket} -> {:noreply, socket}
-      {:error, socket} -> {:noreply, socket}
-    end
-  end
-
-  def handle_event("cancel_door", _params, socket) do
-    {:noreply, assign(socket, :editing, nil) |> assign(:changeset, nil)}
-  end
-
-  # Map interaction events
-  def handle_event("zoom_in", _params, socket) do
-    {:noreply, assign(socket, :zoom, min(socket.assigns.zoom * 1.2, 5.0))}
-  end
-
-  def handle_event("zoom_out", _params, socket) do
-    {:noreply, assign(socket, :zoom, max(socket.assigns.zoom / 1.2, 0.2))}
-  end
-
-  def handle_event("reset_view", _params, socket) do
-    {:noreply,
-     socket
-     |> assign(:zoom, 1.0)
-     |> assign(:pan_x, 0)
-     |> assign(:pan_y, 0)}
-  end
-
-  def handle_event("mousedown", %{"clientX" => x, "clientY" => y}, socket) do
-    {:noreply, assign(socket, :drag_start, %{x: x, y: y})}
-  end
-
-  def handle_event("mousemove", %{"clientX" => x, "clientY" => y}, socket) do
-    case socket.assigns.drag_start do
-      nil ->
-        {:noreply, socket}
-
-      start ->
-        delta_x = x - start.x
-        delta_y = y - start.y
-
-        {:noreply,
-         socket
-         |> assign(:pan_x, socket.assigns.pan_x + delta_x)
-         |> assign(:pan_y, socket.assigns.pan_y + delta_y)
-         |> assign(:drag_start, %{x: x, y: y})}
-    end
-  end
-
-  def handle_event("mouseup", _params, socket) do
-    {:noreply, assign(socket, :drag_start, nil)}
-  end
-
-  def handle_event("mouseleave", _params, socket) do
-    {:noreply, assign(socket, :drag_start, nil)}
-  end
-
-  # Generate default map
-  def handle_event("generate_default_map", _params, socket) do
-    Shard.Repo.delete_all(Door)
-    Shard.Repo.delete_all(Room)
-
-    rooms =
-      for x <- 0..2, y <- 0..2 do
-        name = "Room #{x},#{y}"
-        description = "A room in the default map at coordinates (#{x}, #{y})"
-        room_type = if x == 1 and y == 1, do: "safe_zone", else: "standard"
-
-        {:ok, room} =
-          Map.create_room(%{
-            name: name,
-            description: description,
-            x_coordinate: x,
-            y_coordinate: y,
-            z_coordinate: 0,
-            room_type: room_type,
-            is_public: true
-          })
-
-        room
-      end
-
-    for x <- 0..2, y <- 0..2 do
-      current_room = Enum.find(rooms, &(&1.x_coordinate == x && &1.y_coordinate == y))
-
-      if x < 2 do
-        east_room = Enum.find(rooms, &(&1.x_coordinate == x + 1 && &1.y_coordinate == y))
-
-        Map.create_door(%{
-          from_room_id: current_room.id,
-          to_room_id: east_room.id,
-          direction: "east",
-          door_type: "standard",
-          is_locked: false
-        })
-      end
-
-      if y < 2 do
-        south_room = Enum.find(rooms, &(&1.x_coordinate == x && &1.y_coordinate == y + 1))
-
-        Map.create_door(%{
-          from_room_id: current_room.id,
-          to_room_id: south_room.id,
-          direction: "south",
-          door_type: "standard",
-          is_locked: false
-        })
-      end
-    end
-
-    rooms = Map.list_rooms()
-    doors = Map.list_doors()
-
-    {:noreply,
-     socket
-     |> assign(:rooms, rooms)
-     |> assign(:doors, doors)
-     |> put_flash(:info, "Default 3x3 map generated successfully!")}
-  end
-
-  defp save_room(socket, room_params) do
-    case socket.assigns.editing do
-      :room
-      when not is_nil(socket.assigns.changeset) and
-             not is_nil(socket.assigns.changeset.data.id) ->
-        case Map.update_room(socket.assigns.changeset.data, room_params) do
-          {:ok, _room} ->
-            rooms = Map.list_rooms()
-
-            {:ok,
-             assign(socket, :rooms, rooms)
-             |> assign(:editing, nil)
-             |> assign(:changeset, nil)
-             |> put_flash(:info, "Room updated successfully")}
-
-          {:error, changeset} ->
-            {:error, assign(socket, :changeset, changeset)}
-        end
-
-      _ ->
-        case Map.create_room(room_params) do
-          {:ok, _room} ->
-            rooms = Map.list_rooms()
-
-            {:ok,
-             assign(socket, :rooms, rooms)
-             |> assign(:editing, nil)
-             |> assign(:changeset, nil)
-             |> put_flash(:info, "Room created successfully")}
-
-          {:error, changeset} ->
-            {:error, assign(socket, :changeset, changeset)}
-        end
-    end
-  end
-
-  defp save_door(socket, door_params) do
-    case socket.assigns.editing do
-      :door
-      when not is_nil(socket.assigns.changeset) and
-             not is_nil(socket.assigns.changeset.data.id) ->
-        case Map.update_door(socket.assigns.changeset.data, door_params) do
-          {:ok, _door} ->
-            doors = Map.list_doors()
-
-            {:ok,
-             assign(socket, :doors, doors)
-             |> assign(:editing, nil)
-             |> assign(:changeset, nil)
-             |> put_flash(:info, "Door updated successfully")}
-
-          {:error, changeset} ->
-            {:error, assign(socket, :changeset, changeset)}
-        end
-
-      _ ->
-        case Map.create_door(door_params) do
-          {:ok, _door} ->
-            doors = Map.list_doors()
-
-            {:ok,
-             assign(socket, :doors, doors)
-             |> assign(:editing, nil)
-             |> assign(:changeset, nil)
-             |> put_flash(:info, "Door created successfully")}
-
-          {:error, changeset} ->
-            {:error, assign(socket, :changeset, changeset)}
-        end
-    end
   end
 
   @impl true
@@ -572,339 +233,66 @@ defmodule ShardWeb.AdminLive.Map do
     """
   end
 
-  defp rooms_tab(assigns) do
-    ~H"""
-    <div class="overflow-x-auto">
-      <div class="mb-4">
-        <.button phx-click="new_room" class="btn btn-primary">New Room</.button>
-      </div>
+  # Event handler delegations
+  @impl true
+  def handle_event("change_tab", params, socket),
+    do: MapHandlers.handle_change_tab(params, socket)
 
-      <%= if Enum.empty?(@rooms) do %>
-        <div class="text-center py-8">
-          <p class="text-gray-500">No rooms found.</p>
-        </div>
-      <% else %>
-        <table class="table table-zebra">
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Coordinates</th>
-              <th>Type</th>
-              <th>Public</th>
-              <th>Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            <%= for room <- @rooms do %>
-              <tr>
-                <td>{room.name}</td>
-                <td>({room.x_coordinate}, {room.y_coordinate}, {room.z_coordinate})</td>
-                <td>{room.room_type}</td>
-                <td>{if room.is_public, do: "Yes", else: "No"}</td>
-                <td class="flex space-x-2">
-                  <.button phx-click="view_room" phx-value-id={room.id} class="btn btn-sm btn-info">
-                    View
-                  </.button>
-                  <.button phx-click="edit_room" phx-value-id={room.id} class="btn btn-sm btn-primary">
-                    Edit
-                  </.button>
-                  <.button
-                    phx-click="delete_room"
-                    phx-value-id={room.id}
-                    class="btn btn-sm btn-error"
-                    data-confirm="Are you sure?"
-                  >
-                    Delete
-                  </.button>
-                </td>
-              </tr>
-            <% end %>
-          </tbody>
-        </table>
-      <% end %>
-    </div>
-    """
-  end
+  # Room events
+  def handle_event("new_room", params, socket), do: MapHandlers.handle_new_room(params, socket)
+  def handle_event("edit_room", params, socket), do: MapHandlers.handle_edit_room(params, socket)
+  def handle_event("view_room", params, socket), do: MapHandlers.handle_view_room(params, socket)
 
-  defp doors_tab(assigns) do
-    ~H"""
-    <div class="overflow-x-auto">
-      <div class="mb-4">
-        <.button phx-click="new_door" class="btn btn-primary">New Door</.button>
-      </div>
+  def handle_event("delete_room", params, socket),
+    do: MapHandlers.handle_delete_room(params, socket)
 
-      <%= if Enum.empty?(@doors) do %>
-        <div class="text-center py-8">
-          <p class="text-gray-500">No doors found.</p>
-        </div>
-      <% else %>
-        <table class="table table-zebra">
-          <thead>
-            <tr>
-              <th>From Room</th>
-              <th>To Room</th>
-              <th>Direction</th>
-              <th>Type</th>
-              <th>Locked</th>
-              <th>Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            <%= for door <- @doors do %>
-              <% from_room = Enum.find(@rooms, &(&1.id == door.from_room_id)) %>
-              <% to_room = Enum.find(@rooms, &(&1.id == door.to_room_id)) %>
-              <tr>
-                <td>{if from_room, do: from_room.name, else: "Unknown"}</td>
-                <td>{if to_room, do: to_room.name, else: "Unknown"}</td>
-                <td>{door.direction}</td>
-                <td>{door.door_type}</td>
-                <td>{if door.is_locked, do: "Yes", else: "No"}</td>
-                <td class="flex space-x-2">
-                  <.button phx-click="edit_door" phx-value-id={door.id} class="btn btn-sm btn-primary">
-                    Edit
-                  </.button>
-                  <.button
-                    phx-click="delete_door"
-                    phx-value-id={door.id}
-                    class="btn btn-sm btn-error"
-                    data-confirm="Are you sure?"
-                  >
-                    Delete
-                  </.button>
-                </td>
-              </tr>
-            <% end %>
-          </tbody>
-        </table>
-      <% end %>
-    </div>
-    """
-  end
+  def handle_event("validate_room", params, socket),
+    do: MapHandlers.handle_validate_room(params, socket)
 
-  defp map_visualization(assigns) do
-    ~H"""
-    <div class="bg-base-200 p-6 rounded-lg">
-      <div class="flex justify-between items-center mb-4">
-        <h3 class="text-xl font-bold">Map Visualization</h3>
-        <div class="flex space-x-2">
-          <.button phx-click="zoom_in" class="btn btn-sm">Zoom In</.button>
-          <.button phx-click="zoom_out" class="btn btn-sm">Zoom Out</.button>
-          <.button phx-click="reset_view" class="btn btn-sm">Reset View</.button>
-        </div>
-      </div>
+  def handle_event("save_room", params, socket), do: MapHandlers.handle_save_room(params, socket)
 
-      <%= if Enum.empty?(@rooms) do %>
-        <div class="text-center py-8">
-          <p class="text-gray-500">No rooms available to display.</p>
-        </div>
-      <% else %>
-        <div
-          class="relative overflow-hidden border border-base-300 rounded bg-base-100 cursor-move"
-          style={"height: 600px; transform: scale(#{@zoom}); transform-origin: 0 0;"}
-          id="map-container"
-          phx-hook="MapDrag"
-        >
-          <div
-            class="absolute inset-0"
-            style={"transform: translate(#{@pan_x}px, #{@pan_y}px);"}
-            id="map-content"
-          >
-            <%= for door <- @doors do %>
-              <% from_room = Enum.find(@rooms, &(&1.id == door.from_room_id)) %>
-              <% to_room = Enum.find(@rooms, &(&1.id == door.to_room_id)) %>
-              <%= if from_room && to_room do %>
-                <.door_connection from={from_room} to={to_room} />
-              <% end %>
-            <% end %>
+  def handle_event("apply_and_save", params, socket),
+    do: MapHandlers.handle_apply_and_save(params, socket)
 
-            <%= for room <- @rooms do %>
-              <div
-                class={"absolute rounded border-2 flex items-center justify-center text-center p-1 #{room_classes(room)}"}
-                style={"left: #{room.x_coordinate * 100}px; top: #{room.y_coordinate * 100}px; width: 80px; height: 80px;"}
-              >
-                <div>
-                  <div class="font-bold text-xs truncate w-full">{room.name}</div>
-                  <div class="text-xs mt-1">({room.x_coordinate}, {room.y_coordinate})</div>
-                </div>
-              </div>
-            <% end %>
-          </div>
-        </div>
+  def handle_event("generate_description", params, socket),
+    do: MapHandlers.handle_generate_description(params, socket)
 
-        <div class="mt-4 text-sm text-base-content">
-          <p>Zoom: {Float.round(@zoom, 2)}x | Pan: ({@pan_x}, {@pan_y})</p>
-          <p class="mt-2">Rooms: {Enum.count(@rooms)} | Doors: {Enum.count(@doors)}</p>
-        </div>
-      <% end %>
-    </div>
-    """
-  end
+  def handle_event("cancel_room", params, socket),
+    do: MapHandlers.handle_cancel_room(params, socket)
 
-  defp room_details_tab(assigns) do
-    ~H"""
-    <div class="bg-base-200 p-6 rounded-lg">
-      <div class="flex justify-between items-center mb-4">
-        <h3 class="text-xl font-bold">Room Details: {@room.name}</h3>
-        <.button phx-click="back_to_rooms" class="btn btn-secondary">Back to Rooms</.button>
-      </div>
+  def handle_event("back_to_rooms", params, socket),
+    do: MapHandlers.handle_back_to_rooms(params, socket)
 
-      <.simple_form
-        :let={f}
-        for={@changeset}
-        id="room-details-form"
-        phx-submit="apply_and_save"
-      >
-        <.input field={f[:name]} type="text" label="Name" required />
-        <div class="flex items-end space-x-2">
-          <div class="flex-grow">
-            <.input field={f[:description]} type="textarea" label="Description" />
-          </div>
-          <.button phx-click="generate_description" class="btn btn-secondary" type="button">
-            ✨ Generate with AI
-          </.button>
-        </div>
-        <div class="grid grid-cols-3 gap-4">
-          <.input field={f[:x_coordinate]} type="number" label="X Coordinate" />
-          <.input field={f[:y_coordinate]} type="number" label="Y Coordinate" />
-          <.input field={f[:z_coordinate]} type="number" label="Z Coordinate" />
-        </div>
-        <.input
-          field={f[:room_type]}
-          type="select"
-          label="Type"
-          prompt="Choose a type"
-          options={[
-            {"Standard", "standard"},
-            {"Safe Zone", "safe_zone"},
-            {"Shop", "shop"},
-            {"Dungeon", "dungeon"},
-            {"Treasure Room", "treasure_room"},
-            {"Trap Room", "trap_room"}
-          ]}
-        />
-        <.input field={f[:is_public]} type="checkbox" label="Public Room" />
+  # Door events
+  def handle_event("new_door", params, socket), do: MapHandlers.handle_new_door(params, socket)
+  def handle_event("edit_door", params, socket), do: MapHandlers.handle_edit_door(params, socket)
 
-        <div class="mt-6">
-          <h4 class="text-lg font-bold mb-2">Doors Leading From This Room</h4>
-          <%= if Enum.empty?(@doors_from) do %>
-            <p class="text-gray-500">No doors lead from this room.</p>
-          <% else %>
-            <table class="table table-zebra">
-              <thead>
-                <tr>
-                  <th>To Room</th>
-                  <th>Direction</th>
-                  <th>Type</th>
-                  <th>Locked</th>
-                </tr>
-              </thead>
-              <tbody>
-                <%= for door <- @doors_from do %>
-                  <tr>
-                    <td>{door.to_room.name}</td>
-                    <td>{door.direction}</td>
-                    <td>{door.door_type}</td>
-                    <td>{if door.is_locked, do: "Yes", else: "No"}</td>
-                  </tr>
-                <% end %>
-              </tbody>
-            </table>
-          <% end %>
-        </div>
+  def handle_event("delete_door", params, socket),
+    do: MapHandlers.handle_delete_door(params, socket)
 
-        <div class="mt-6">
-          <h4 class="text-lg font-bold mb-2">Doors Leading To This Room</h4>
-          <%= if Enum.empty?(@doors_to) do %>
-            <p class="text-gray-500">No doors lead to this room.</p>
-          <% else %>
-            <table class="table table-zebra">
-              <thead>
-                <tr>
-                  <th>From Room</th>
-                  <th>Direction</th>
-                  <th>Type</th>
-                  <th>Locked</th>
-                </tr>
-              </thead>
-              <tbody>
-                <%= for door <- @doors_to do %>
-                  <tr>
-                    <td>{door.from_room.name}</td>
-                    <td>{door.direction}</td>
-                    <td>{door.door_type}</td>
-                    <td>{if door.is_locked, do: "Yes", else: "No"}</td>
-                  </tr>
-                <% end %>
-              </tbody>
-            </table>
-          <% end %>
-        </div>
+  def handle_event("validate_door", params, socket),
+    do: MapHandlers.handle_validate_door(params, socket)
 
-        <:actions>
-          <.button phx-disable-with="Saving..." class="btn btn-primary">Apply and Save</.button>
-        </:actions>
-      </.simple_form>
-    </div>
-    """
-  end
+  def handle_event("save_door", params, socket), do: MapHandlers.handle_save_door(params, socket)
 
-  defp door_connection(assigns) do
-    from_x = assigns.from.x_coordinate * 100 + 40
-    from_y = assigns.from.y_coordinate * 100 + 40
-    to_x = assigns.to.x_coordinate * 100 + 40
-    to_y = assigns.to.y_coordinate * 100 + 40
+  def handle_event("cancel_door", params, socket),
+    do: MapHandlers.handle_cancel_door(params, socket)
 
-    assigns
-    |> assign(:from_x, from_x)
-    |> assign(:from_y, from_y)
-    |> assign(:to_x, to_x)
-    |> assign(:to_y, to_y)
-    |> render_door_connection()
-  end
+  # Map interaction events
+  def handle_event("zoom_in", params, socket), do: MapHandlers.handle_zoom_in(params, socket)
+  def handle_event("zoom_out", params, socket), do: MapHandlers.handle_zoom_out(params, socket)
 
-  defp render_door_connection(assigns) do
-    ~H"""
-    <svg class="absolute inset-0 w-full h-full pointer-events-none">
-      <line
-        x1={@from_x}
-        y1={@from_y}
-        x2={@to_x}
-        y2={@to_y}
-        stroke="currentColor"
-        stroke-width="2"
-        stroke-dasharray="5,5"
-        class="text-base-content/30"
-      />
-    </svg>
-    """
-  end
+  def handle_event("reset_view", params, socket),
+    do: MapHandlers.handle_reset_view(params, socket)
 
-  defp room_classes(room) do
-    base_classes = "flex flex-col items-center justify-center text-xs"
+  def handle_event("mousedown", params, socket), do: MapHandlers.handle_mousedown(params, socket)
+  def handle_event("mousemove", params, socket), do: MapHandlers.handle_mousemove(params, socket)
+  def handle_event("mouseup", params, socket), do: MapHandlers.handle_mouseup(params, socket)
 
-    type_classes =
-      case room.room_type do
-        "safe_zone" ->
-          "bg-green-200 border-green-600 dark:bg-green-900/30 dark:border-green-500"
+  def handle_event("mouseleave", params, socket),
+    do: MapHandlers.handle_mouseleave(params, socket)
 
-        "shop" ->
-          "bg-blue-200 border-blue-600 dark:bg-blue-900/30 dark:border-blue-500"
-
-        "dungeon" ->
-          "bg-red-200 border-red-600 dark:bg-red-900/30 dark:border-red-500"
-
-        "treasure_room" ->
-          "bg-yellow-200 border-yellow-600 dark:bg-yellow-900/30 dark:border-yellow-500"
-
-        "trap_room" ->
-          "bg-pink-200 border-pink-600 dark:bg-pink-900/30 dark:border-pink-500"
-
-        _ ->
-          "bg-gray-200 border-gray-600 dark:bg-gray-700/30 dark:border-gray-500"
-      end
-
-    text_classes = "text-gray-800 dark:text-gray-200"
-
-    "#{base_classes} #{type_classes} #{text_classes}"
-  end
+  # Generate default map
+  def handle_event("generate_default_map", params, socket),
+    do: MapHandlers.handle_generate_default_map(params, socket)
 end
